feat(cache): improve cache control message
